@@ -10,15 +10,16 @@ return (
            <h1> About Me</h1>
            </div> 
     <div className= "aboutMe">
-    <h2 id="aboutMeText" style={{fontWeight:"350" }}>
+    <h2 style={{fontWeight:"350" }}>
          With a history in education, creating and running my own business, and a lifetime of studying music and the arts, I am a passionate and dedicated learner who believes in using the creative mind to find solutions and solve any problem that comes my way. 
         
         <br/>
         <br/>
-        Playing music for most of my life has taught the values of collaboration, and whether it is working with my students, playing jazz guitar, or learning different programming languages, I believe that with time and dedicated practice, anything is achievable. 
+        Playing in bands for most of my life has taught the values of collaboration, and whether it is working with my students, playing jazz guitar, or learning different programming languages, I believe that with time and dedicated practice, anything is achievable. 
         <br/>
         <br/>
         <div className = "pic-container">
+            
     <img src={ProfPic} className = "prof-pic" alt="prof-picture" />
     </div>
 
